Scoa/requirements.txt
python-dotenv
pymongo
flask
requests
python-telegram-bot==13.15